Add explanatory comments & use CompilerSet
I'm not sure of when to use let vs set vs setlocal but most compiler
files seem to define (if needed) and then use CompilerSet this way.

Setting the rules as strings is easier to read as you don't have to
escape the spaces but I'm not sure how to get it working with
CompilerSet so we just go with it. It is probably quite basic but I get
confused.