Add 'compiler' command to ftplugin
When trying to get a set up working from fresh, I couldn't get the
compiler ':make' functionality to work. It seems it isn't picked up
automatically and needs a line like this.

We also move the augroup to the bottom and give an END.

And add a check at the start to make sure we're only sourcing this file
once and only if we've not loaded any other ftplugins.

These changes are entirely based on the ftplugin/rust.vim:

   https://github.com/rust-lang/rust.vim/blob/master/ftplugin/rust.vim

There is a chance that we're copying bad practices from them but it
seems reasonable. I don't know anything about augroups but it seems like
they should have an end and moving it to the bottom makes it clearer
what is and isn't intended to be included in the group.